Project/Sample Reference
CMakeLists.txt- Added directories to get the needed files for the mmu
prj.conf- added the needed stuff to get the Kconfig options for MMU working. The HEAP_MEM_POOL_SIZE was needed to get k_malloc to be included
main.c- changed the name of the function to reflect the riscv_mmu.h function
riscv_mmu.h- added in the define for PTE_USER & the map_page declaration so it wasn't implicit in the main.c. Probably still need to add the other functions there as well.
mmu.c- fixed the type cast warnings and added the flag checks for the entries